requirements.txt
scikit-learn==1.4.0
transformers==4.37.2
torch==2.2.0
numpy==1.26.4
python-docx==1.1.0
pandas==2.2.0
pinecone-client==3.0.2
streamlit==1.31.1
jsonschema==4.21.1
openai==1.12.0
streamlit-scrollable-textbox==0.0.3